a little test re-org
