+++ JsonSchema.CodeGeneration.Tests/Usings.cs
@@ -1,1 +1,3 @@
 global using NUnit.Framework;
+
+global using static Json.Schema.CodeGeneration.Tests.AssertHelpers;
